refactor: move `mark/mixins` and `mark/valueref` to `mark/encode/*` and split files for clarity
This should allow us to read the mark compiler code easier to read
