+++ src/compile/guide.ts
@@ -2,7 +2,7 @@ import {ValueDef} from '../channeldef';
 import {GuideEncodingEntry} from '../guide';
 import {keys} from '../util';
 import {VgEncodeChannel} from '../vega.schema';
-import {wrapCondition} from './mark/mixins';
+import {wrapCondition} from './mark/encode';
 import {UnitModel} from './unit';
 
 export function guideEncodeEntry(encoding: GuideEncodingEntry, model: UnitModel) {
