+++ src/compile/mark/area.ts
@@ -1,29 +1,29 @@
 import {UnitModel} from '../unit';
 import {MarkCompiler} from './base';
-import * as mixins from './mixins';
+import * as encode from './encode';
 
 export const area: MarkCompiler = {
   vgMark: 'area',
   encodeEntry: (model: UnitModel) => {
     return {
-      ...mixins.baseEncodeEntry(model, {
+      ...encode.baseEncodeEntry(model, {
         align: 'ignore',
         baseline: 'ignore',
         color: 'include',
         orient: 'include',
         size: 'ignore'
       }),
-      ...mixins.pointOrRangePosition('x', model, {
-        defaultRef: 'zeroOrMin',
-        defaultRef2: 'zeroOrMin',
+      ...encode.pointOrRangePosition('x', model, {
+        defaultPos: 'zeroOrMin',
+        defaultPos2: 'zeroOrMin',
         range: model.markDef.orient === 'horizontal'
       }),
-      ...mixins.pointOrRangePosition('y', model, {
-        defaultRef: 'zeroOrMin',
-        defaultRef2: 'zeroOrMin',
+      ...encode.pointOrRangePosition('y', model, {
+        defaultPos: 'zeroOrMin',
+        defaultPos2: 'zeroOrMin',
         range: model.markDef.orient === 'vertical'
       }),
-      ...mixins.defined(model)
+      ...encode.defined(model)
     };
   }
 };
