+++ src/compile/mark/bar.ts
@@ -1,21 +1,20 @@
 import {UnitModel} from '../unit';
 import {MarkCompiler} from './base';
-import * as mixins from './mixins';
-import {rectPosition} from './rect';
+import * as encode from './encode';
 
 export const bar: MarkCompiler = {
   vgMark: 'rect',
   encodeEntry: (model: UnitModel) => {
     return {
-      ...mixins.baseEncodeEntry(model, {
+      ...encode.baseEncodeEntry(model, {
         align: 'ignore',
         baseline: 'ignore',
         color: 'include',
         orient: 'ignore',
         size: 'ignore'
       }),
-      ...rectPosition(model, 'x', 'bar'),
-      ...rectPosition(model, 'y', 'bar')
+      ...encode.rectPosition(model, 'x', 'bar'),
+      ...encode.rectPosition(model, 'y', 'bar')
     };
   }
 };
