+++ src/compile/mark/image.ts
@@ -1,22 +1,21 @@
 import {UnitModel} from '../unit';
 import {MarkCompiler} from './base';
-import * as mixins from './mixins';
-import {rectPosition} from './rect';
+import * as encode from './encode';
 
 export const image: MarkCompiler = {
   vgMark: 'image',
   encodeEntry: (model: UnitModel) => {
     return {
-      ...mixins.baseEncodeEntry(model, {
+      ...encode.baseEncodeEntry(model, {
         align: 'ignore',
         baseline: 'ignore',
         color: 'ignore',
         orient: 'ignore',
         size: 'ignore'
       }),
-      ...rectPosition(model, 'x', 'image'),
-      ...rectPosition(model, 'y', 'image'),
-      ...mixins.text(model, 'url')
+      ...encode.rectPosition(model, 'x', 'image'),
+      ...encode.rectPosition(model, 'y', 'image'),
+      ...encode.text(model, 'url')
     };
   }
 };
