+++ src/compile/mark/line.ts
@@ -1,27 +1,24 @@
 import {UnitModel} from '../unit';
 import {MarkCompiler} from './base';
-import * as mixins from './mixins';
-import * as ref from './valueref';
+import * as encode from './encode';
 
 export const line: MarkCompiler = {
   vgMark: 'line',
   encodeEntry: (model: UnitModel) => {
-    const {width, height} = model;
-
     return {
-      ...mixins.baseEncodeEntry(model, {
+      ...encode.baseEncodeEntry(model, {
         align: 'ignore',
         baseline: 'ignore',
         color: 'include',
         size: 'ignore',
         orient: 'ignore'
       }),
-      ...mixins.pointPosition('x', model, ref.mid(width)),
-      ...mixins.pointPosition('y', model, ref.mid(height)),
-      ...mixins.nonPosition('size', model, {
+      ...encode.pointPosition('x', model, {defaultPos: 'mid'}),
+      ...encode.pointPosition('y', model, {defaultPos: 'mid'}),
+      ...encode.nonPosition('size', model, {
         vgChannel: 'strokeWidth' // VL's line size is strokeWidth
       }),
-      ...mixins.defined(model)
+      ...encode.defined(model)
     };
   }
 };
@@ -29,20 +26,18 @@ export const line: MarkCompiler = {
 export const trail: MarkCompiler = {
   vgMark: 'trail',
   encodeEntry: (model: UnitModel) => {
-    const {width, height} = model;
-
     return {
-      ...mixins.baseEncodeEntry(model, {
+      ...encode.baseEncodeEntry(model, {
         align: 'ignore',
         baseline: 'ignore',
         color: 'include',
         size: 'include',
         orient: 'ignore'
       }),
-      ...mixins.pointPosition('x', model, ref.mid(width)),
-      ...mixins.pointPosition('y', model, ref.mid(height)),
-      ...mixins.nonPosition('size', model),
-      ...mixins.defined(model)
+      ...encode.pointPosition('x', model, {defaultPos: 'mid'}),
+      ...encode.pointPosition('y', model, {defaultPos: 'mid'}),
+      ...encode.nonPosition('size', model),
+      ...encode.defined(model)
     };
   }
 };
