+++ src/compile/mark/text.ts
@@ -4,31 +4,30 @@ import {MarkDef} from '../../mark';
 import {getMarkConfig} from '../common';
 import {UnitModel} from '../unit';
 import {MarkCompiler} from './base';
-import * as mixins from './mixins';
-import * as ref from './valueref';
+import * as encode from './encode';
 
 export const text: MarkCompiler = {
   vgMark: 'text',
 
   encodeEntry: (model: UnitModel) => {
-    const {config, encoding, width, height} = model;
+    const {config, encoding} = model;
 
     return {
-      ...mixins.baseEncodeEntry(model, {
+      ...encode.baseEncodeEntry(model, {
         align: 'include',
         baseline: 'include',
         color: 'include',
         size: 'ignore',
         orient: 'ignore'
       }),
-      ...mixins.pointPosition('x', model, ref.mid(width)),
-      ...mixins.pointPosition('y', model, ref.mid(height)),
-      ...mixins.text(model),
-      ...mixins.nonPosition('size', model, {
+      ...encode.pointPosition('x', model, {defaultPos: 'mid'}),
+      ...encode.pointPosition('y', model, {defaultPos: 'mid'}),
+      ...encode.text(model),
+      ...encode.nonPosition('size', model, {
         vgChannel: 'fontSize' // VL's text size is fontSize
       }),
-      ...mixins.valueIfDefined('align', align(model.markDef, encoding, config)),
-      ...mixins.valueIfDefined('baseline', baseline(model.markDef, encoding, config))
+      ...encode.valueIfDefined('align', align(model.markDef, encoding, config)),
+      ...encode.valueIfDefined('baseline', baseline(model.markDef, encoding, config))
     };
   }
 };
